Update SplinesUtils.R
Regress splines in a linear mixed model of "lme" class (using nlme package) can now be exported as PiecePoly. This enhancement in functionality unfortunately resulted in structural change of RegBsplineAsPiecePoly function. This function is now obsolete and superseded by RegSplineAsPiecePoly. However, it may still be used for now, with a pop-up warning.